Overhaul application
Drop IntelliJ junk. Integrate with SBT. Add a dependency on Scallop, and
implement a proper option parser. Flatten the Star case class. Add unit
tests. Add a dependency on the Hadoop client libraries, and make it
possible to write files to anywhere Hadoop can (including the local
filesystem and HDFS).